jetson-hcv-risk-poc/edge/requirements.txt
# Phase 0 edge (Jetson / dev workstation)
opencv-python-headless>=4.8.0
pyserial>=3.5
PyYAML>=6.0
